LayoutTests/imported/w3c/web-platform-tests/css/css-tables/tentative/table-height-redistribution-expected.txt

Tests for redistribution of table height to row group height.

This algorithm has not been standardized. Browsers implementations disagree a lot. Legacy Chrome used to always distribute all the height to the first tbody.
Major incompatibility: Legacy:
- ignores any height set on sections.
- does not size the table unless section has a TR
- does not grow sections without TDs
- only distributes height to the 1st tbody
FF
- does not prioritize thead for height distribution most of the time.
- y offset of multiple tbodies can be incorrect.
Empty table

Empty tables always grow to specified height in all browsers.
no sections
no sections, no border spacing
collapsed table no sections
fixed table no sections
One TBODY

The big difference here is between empty TBODY, and a body with an empty TR
FF: always sizes the table, only sizes TBODY if it has TR.
Legacy: does not size the table unless TBODY has TR
FF: sizes the table, but tbody size remains 0 unless tbody is not empty.
Legacy table size remains 0, or border-spacing
single empty tbody
b:0
single tbody+tr
b:100
border spacing
FF/Legacy do not apply border-spacing
b:100
collapsed table
b:100
fixed table
b:100
tbody has fixed height
FF adds tbody height to table?
b:100
tbody has fixed height > table
Legacy: table size wins. FF: table size wins, but body grows to 300px?
b:100
tr has fixed height > table
FF: table size wins, but body is 300.
b:200
tbody has percentage height > table
b:100
tr has percentage height > table
FF/Legacy: table wins. FF: body is 200px
b:100
non-empty tbody
b:100
non-empty thead
h:0
THEAD TBODY

empty thead, empty tbody
FF thead/tbody both grow
h:0
b:100
sized thead, empty tbody
h:20
b:80
thead
table layout fixed, thead with td, tbody
h:20
b:80
thead
x
table layout fixed, thead+td, tbody+td
h:20
b:80
thead
x
thead with td
h:20
b:80
thead
tfoot with td
b:80
f:20
tfoot
Multiple TBODY

Legacy does not distribute any heights when tr is empty.
Legacy always distributes everything to 1st tbody.
2 tbody
b:0
b:0
2 tbody, with non-empty tds
Legacy distributes everything to 1st tbody
b:82
b:18
x
x
2 tbody, 40%, auto, no td
FF: distributes everything to auto when empty. Legacy does not distribute
b:0
b:0
2 tbody, 40%, auto, non-empty td
b:82
b:18
x
x
2 tbody, 40px, auto, non-empty td
FF gets confused with 2nd body placement
b:82
b:18
x
x
2 tbody, 40%, 40px, non-empty td
FF splits evenly
b:82
b:18
x
x
Sized THEAD/TBODY

FF does not prioritize TBODY for distribution
20px thead, 30px tbody
h:18
b:82
x
x
20px thead's tr, 30px tbody's tr
h:20
b:80
x
x
20px thead's td, 30px tbody's td
h:20
b:80
x
x

PASS table 1
PASS table 2
PASS table 3
PASS table 4
FAIL table 5 assert_equals:
<table data-expected-height="100">
  <tbody data-expected-height="100"></tbody>
</table>
height expected 100 but got 0
PASS table 6
FAIL table 7 assert_equals:
<table style="border-spacing: 10px" data-expected-height="100">
  <tbody data-expected-height="80"><tr></tr></tbody>
</table>
height expected 80 but got 100
PASS table 8
PASS table 9
PASS table 10
FAIL table 11 assert_equals:
<table data-expected-height="200">
  <tbody style="height:200px" data-expected-height="200"><tr></tr></tbody>
</table>
height expected 200 but got 100
PASS table 12
PASS table 13
PASS table 14
PASS table 15
FAIL table 16 assert_equals:
<table data-expected-height="100">
  <thead data-expected-height="100"><tr></tr></thead>
</table>
height expected 100 but got 0
PASS table 17
PASS table 18
PASS table 19
PASS table 20
PASS table 21
PASS table 22
FAIL table 23 assert_equals:
<table data-expected-height="100">
  <tbody data-expected-height="50"><tr></tr></tbody>
  <tbody data-expected-height="50"><tr></tr></tbody>
</table>
height expected 100 but got 0
FAIL table 24 assert_equals:
<table data-expected-height="100">
  <tbody data-expected-height="50"><tr><td>x</td></tr></tbody>
  <tbody data-expected-height="50"><tr><td>x</td></tr></tbody>
</table>
height expected 50 but got 82
FAIL table 25 assert_equals:
<table data-expected-height="100">
  <tbody style="height:40%" data-expected-height="40"><tr></tr></tbody>
  <tbody data-expected-height="60"><tr></tr></tbody>
</table>
height expected 100 but got 0
FAIL table 26 assert_equals:
<table data-expected-height="100">
  <tbody style="height:40%" data-expected-height="40"><tr><td>x</td></tr></tbody>
  <tbody data-expected-height="60"><tr><td>x</td></tr></tbody>
</table>
height expected 40 but got 82
FAIL table 27 assert_equals:
<table data-expected-height="100">
  <tbody style="height:40px" data-expected-height="40"><tr><td>x</td></tr></tbody>
  <tbody data-expected-height="60" data-offset-y="40"><tr><td>x</td></tr></tbody>
</table>
height expected 40 but got 82
FAIL table 28 assert_equals:
<table data-expected-height="100">
  <tbody style="height:40%" data-expected-height="40"><tr><td>x</td></tr></tbody>
  <tbody style="height:40px" data-expected-height="60" data-offset-y="40"><tr><td>x</td></tr></tbody>
</table>
height expected 40 but got 82
FAIL table 29 assert_equals:
<table data-expected-height="100">
  <thead data-expected-height="20" style="height:20px">
    <tr><td>x</td></tr>
  </thead>
  <tbody data-expected-height="80" style="height:30px">
    <tr><td>x</td></tr>
  </tbody>
</table>
height expected 20 but got 18
PASS table 30
PASS table 31
